Don't spawn unnecessary goroutines for StaleWhileRevalidate
Also moved racy tests into a separate file that doesn't get build when
the race detector is enabled. These tests are racy because they don't
use the mutexes in monitorFunc to read the values.
